C++ change of Encoder to QuadratureEncoder
fixes #2185
